requirements.txt
biopython==1.83
et-xmlfile==1.1.0
numpy==1.26.4
openpyxl==3.1.2
